requirements.txt
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv-python==0.0.1
load-dotenv==0.1.0
loadotenv==1.0.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.1
sqlparse==0.5.2
